Two main files.
MAIN is made for tracking all the transpositions. Able to deal with relarively small number of free spots (about 1-7).

Main but no T doesn't track all the transpositions of 0 and 1. The main purpose is to calculate the number of excited states in a given enegry interval. Able to deal with big numbers (174-20=154 free spots).